requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==45.0.2
fastapi==0.115.12
granian==2.2.6
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msal==1.32.3
packaging==25.0
platformdirs==4.3.8
psutil==7.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
python-decouple==3.8
python-engineio==4.12.1
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
redis==5.3.0
reflex==0.7.11
reflex-hosting-cli==0.1.49
requests==2.32.3
rich==14.0.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
tabulate==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
watchfiles==1.0.5
wrapt==1.17.2
wsproto==1.2.0
